Djangoqr/requirement.txt
asgiref==3.8.1
Django==5.1.4
numpy==2.2.1
opencv-python==4.10.0.84
pillow==11.1.0
pyzbar==0.1.9
qrcode==8.0
sqlparse==0.5.3
